+++ app/Actions/Photo/Strategies/RotateStrategy.php
@@ -5,7 +5,7 @@ namespace App\Actions\Photo\Strategies;
 use App\Contracts\Exceptions\LycheeException;
 use App\Contracts\Models\AbstractSizeVariantNamingStrategy;
 use App\Contracts\Models\SizeVariantFactory;
-use App\DTO\ImageDimension;
+use App\Data\ImageDimension;
 use App\Enum\SizeVariantType;
 use App\Exceptions\Handler;
 use App\Exceptions\Internal\FrameworkException;
