+++ app/Http/Requests/Album/SetAlbumProtectionPolicyRequest.php
@@ -5,7 +5,7 @@ namespace App\Http\Requests\Album;
 use App\Contracts\Http\Requests\HasBaseAlbum;
 use App\Contracts\Http\Requests\HasPassword;
 use App\Contracts\Http\Requests\RequestAttribute;
-use App\DTO\AlbumProtectionPolicy;
+use App\Data\AlbumProtectionPolicy;
 use App\Http\Requests\BaseApiRequest;
 use App\Http\Requests\Traits\Authorize\AuthorizeCanEditAlbumTrait;
 use App\Http\Requests\Traits\HasBaseAlbumTrait;
