+++ app/Http/Requests/Album/SetAlbumSortingRequest.php
@@ -5,7 +5,7 @@ namespace App\Http\Requests\Album;
 use App\Contracts\Http\Requests\HasBaseAlbum;
 use App\Contracts\Http\Requests\HasSortingCriterion;
 use App\Contracts\Http\Requests\RequestAttribute;
-use App\DTO\PhotoSortingCriterion;
+use App\Data\PhotoSortingCriterion;
 use App\Enum\ColumnSortingPhotoType;
 use App\Enum\OrderSortingType;
 use App\Http\Requests\BaseApiRequest;
