+++ app/Image/Handlers/ImagickHandler.php
@@ -7,7 +7,7 @@ namespace App\Image\Handlers;
 use App\Contracts\Image\ImageHandlerInterface;
 use App\Contracts\Image\MediaFile;
 use App\Contracts\Image\StreamStats;
-use App\DTO\ImageDimension;
+use App\Data\ImageDimension;
 use App\Exceptions\ImageProcessingException;
 use App\Exceptions\MediaFileOperationException;
 use App\Image\Files\InMemoryBuffer;
