+++ app/Image/SizeVariantDefaultFactory.php
@@ -6,7 +6,7 @@ use App\Contracts\Exceptions\LycheeException;
 use App\Contracts\Image\ImageHandlerInterface;
 use App\Contracts\Models\AbstractSizeVariantNamingStrategy;
 use App\Contracts\Models\SizeVariantFactory;
-use App\DTO\ImageDimension;
+use App\Data\ImageDimension;
 use App\Enum\SizeVariantType;
 use App\Exceptions\ConfigurationException;
 use App\Exceptions\ExternalComponentMissingException;
