+++ app/Models/Extensions/SizeVariants.php
@@ -3,8 +3,8 @@
 namespace App\Models\Extensions;
 
 use App\Actions\SizeVariant\Delete;
+use App\Data\ImageDimension;
 use App\DTO\AbstractDTO;
-use App\DTO\ImageDimension;
 use App\Enum\SizeVariantType;
 use App\Exceptions\Internal\IllegalOrderOfOperationException;
 use App\Exceptions\Internal\InvalidSizeVariantException;
